project-root/bot/requirements.txt
discord.py>=1.7.3
torch
scikit-learn
pandas
matplotlib
joblib
python-dotenv
sqlalchemy
pymysql